requirements.txt
django==4.0.6
djangorestframework==3.13.1
matplotlib==3.5.2
mysqlclient==2.1.1
numpy==1.23.0
pandas== 1.4.3
drf-yasg==1.21.3
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-cors-headers==3.4.0
django-rest-framework==0.1.0
django-simple-history==2.11.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
et-xmlfile==1.0.1
Faker==13.3.4
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.3
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
psycopg2-binary==2.8.5
PyJWT==2.3.0
pyparsing==2.4.6
python-dateutil==2.8.2
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.14.0
sqlparse==0.4.1
toml==0.10.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1